Experiment around #amqqueue{} backward-compatibility
The #amqqueue{} record is now isolated in the `amqqueue` module: all
accesses and modifications to the record are made through this helper.

`amqqueue` knows about the new #amqqueue{} record introduced with quorum
queues. If the actual record is the old one (as defined in RabbitMQ
3.7.x), it calls `amqqueue_v1` which knows about the old definition.

When it needs to produce an #amqqueue{} record, it verifies if it is
allowed to (currently, this is always true) and proceeds. If an old
format #amqqueue{} is required, the creation is deferred to
`amqqueue_v1`.

The new modules come with a couple headers: they provide macros to
replace pattern matching with `when` conditions. The `amqqueue_v*.hrl`
headers are generated using the `macros()` functions in `amqqueue` and
`amqqueue_v1` modules.

Note: Another possibility would be to (ab)use the `options` record field
which is already a map. New quorum-queue-specific fields could be moved
to new entries in this map.

[#159298729]